backend/requirements.txt
# Основные зависимости
fastapi==0.109.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.11.7
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0
loguru==0.7.2

# Инструменты разработки
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
ruff==0.2.1
black==24.3.0
isort==5.12.0
mypy==1.8.0
pre-commit==3.6.0
bandit==1.7.5
